refactor: replace entity exposure with DTO layer and mapping logic

+++ src/main/resources/db/migration/V2__create_product_table.sql
@@ -3,6 +3,6 @@ CREATE TABLE products.product (
     product_identifier VARCHAR NOT NULL,
     product_name VARCHAR(100) NOT NULL,
     product_description VARCHAR NOT NULL,
-    price FLOAT NOT NULL,
+    price NUMERIC(10, 2) NOT NULL,
     category_id BIGINT REFERENCES products.category(id)
 );
